+++ lib/axn/core/validation/validators/model_validator.rb
@@ -5,10 +5,24 @@ require "active_model"
 module Axn
   module Validators
     class ModelValidator < ActiveModel::EachValidator
-      def check_validity!
-        return unless options[:with].nil?
+      # Syntactic sugar: model: User -> model: { klass: User }
+      def self.apply_syntactic_sugar(value, fields)
+        (value.is_a?(Hash) ? value.dup : { klass: value }).tap do |options|
+          # Set default klass based on field name if not provided
+          options[:klass] ||= fields.first.to_s.classify
 
-        raise ArgumentError, "must supply :with"
+          # Constantize string klass names
+          options[:klass] = options[:klass].constantize if options[:klass].is_a?(String)
+
+          # Set default finder if not provided
+          options[:finder] ||= :find
+        end
+      end
+
+      def check_validity!
+        return unless options[:klass].nil?
+
+        raise ArgumentError, "must supply :klass"
       end
 
       def validate_each(record, attribute, value)